Move `ErrorHandler` and `Position` to the top level syntax package
